requirements.txt
aiohttp
aio_pika
python-dotenv
colorlog
pytz
aiosseclient